+++ one.cpp
@@ -495,7 +495,7 @@ static int cli(int argc,char **argv)
 					(std::string("/network/") + arg1).c_str(),
 					requestHeaders,
 					jsons,
-					strlen(jsons),
+					(unsigned long)strlen(jsons),
 					responseHeaders,
 					responseBody);
 				if (scode == 200) {
@@ -572,7 +572,7 @@ static int idtool(int argc,char **argv)
 		int vanityBits = 0;
 		if (argc >= 5) {
 			vanity = Utils::hexStrToU64(argv[4]) & 0xffffffffffULL;
-			vanityBits = 4 * strlen(argv[4]);
+			vanityBits = 4 * (int)strlen(argv[4]);
 			if (vanityBits > 40)
 				vanityBits = 40;
 		}
